requirements.txt
asgiref==3.4.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2.6
djangorestframework==3.12.4
djongo==1.3.6
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.12.0
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.4
Werkzeug==1.0.1
